backend/requirements.txt
Flask
Flask-CORS
yfinance
pandas
python-dotenv
finnhub-python
scikit-learn
xgboost
statsmodels
numpy
requests
